fix (renderer) : fix lights rendering loops
Fix a bug that prevent lights from being renderer properly when the light count was not equal to the maximum allowed lights.
The light count to render for each king of light is now passed as a uniform parameter to the shader.
The maximum light count for each type of light as been raised to 50.